+++ .vs/launch.vs.json
@@ -1,0 +1,12 @@
+{
+  "version": "0.2.1",
+  "defaults": {},
+  "configurations": [
+    {
+      "type": "native",
+      "name": "build.exe",
+      "project": "build\\build.exe",
+      "projectTarget": ""
+    }
+  ]
+}
