+++ .vs/tasks.vs.json
@@ -1,0 +1,67 @@
+{
+  "version": "0.2.1",
+  "tasks": [
+    {
+      "taskLabel": "bootstrap",
+      "contextType": "build",
+      "appliesTo": "/build/*",
+      "type": "launch",
+      "workingDirectory": "${workspaceRoot}/build/",
+      "commands": [
+        {
+          "command": "${env.COMSPEC}",
+          "args": [
+            "echo compiling build"
+          ]
+        },
+        {
+          "command": "gcc",
+          "args": [
+            "builder.c",
+            "-g",
+            "-ggdb",
+            "-o builder"
+          ]
+        },
+        {
+          "command": "${env.COMSPEC}",
+          "args": [
+            "echo build compile successfull"
+          ]
+        }
+      ]
+    },
+    {
+      "taskLabel": "build",
+      "contextType": "custom",
+      "workingDirectory": "${workspaceRoot}/build/",
+      "appliesTo": "*",
+      "type": "launch",
+      "commands": [
+        {
+          "command": "${env.COMSPEC}",
+          "args": [
+            "echo started build"
+          ]
+        },
+        {
+          "command": "builder",
+          "args": [
+            "BUILD"
+          ]
+        },
+        {
+          "command": "${env.COMSPEC}",
+          "args": [
+            "echo build successfull"
+          ]
+        }
+      ]
+    },
+    {
+      "taskLabel": "task-builder",
+      "appliesTo": "build/builder.exe",
+      "type": "launch"
+    }
+  ]
+}
