--- AvGraphicsUtilities.sln
@@ -1,41 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.5.33530.505
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "GraphicsUtilities", "GraphicsUtilities\GraphicsUtilities.vcxproj", "{586227E2-BC51-4659-A0A9-68559E7CB627}"
-EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "GraphicsUtilitiesTest", "GraphicsUtilitiesTest\GraphicsUtilitiesTest.vcxproj", "{7A352F3E-0C05-4EBA-BD74-BF6EDA787D57}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{586227E2-BC51-4659-A0A9-68559E7CB627}.Debug|x64.ActiveCfg = Debug|x64
-		{586227E2-BC51-4659-A0A9-68559E7CB627}.Debug|x64.Build.0 = Debug|x64
-		{586227E2-BC51-4659-A0A9-68559E7CB627}.Debug|x86.ActiveCfg = Debug|Win32
-		{586227E2-BC51-4659-A0A9-68559E7CB627}.Debug|x86.Build.0 = Debug|Win32
-		{586227E2-BC51-4659-A0A9-68559E7CB627}.Release|x64.ActiveCfg = Release|x64
-		{586227E2-BC51-4659-A0A9-68559E7CB627}.Release|x64.Build.0 = Release|x64
-		{586227E2-BC51-4659-A0A9-68559E7CB627}.Release|x86.ActiveCfg = Release|Win32
-		{586227E2-BC51-4659-A0A9-68559E7CB627}.Release|x86.Build.0 = Release|Win32
-		{7A352F3E-0C05-4EBA-BD74-BF6EDA787D57}.Debug|x64.ActiveCfg = Debug|x64
-		{7A352F3E-0C05-4EBA-BD74-BF6EDA787D57}.Debug|x64.Build.0 = Debug|x64
-		{7A352F3E-0C05-4EBA-BD74-BF6EDA787D57}.Debug|x86.ActiveCfg = Debug|Win32
-		{7A352F3E-0C05-4EBA-BD74-BF6EDA787D57}.Debug|x86.Build.0 = Debug|Win32
-		{7A352F3E-0C05-4EBA-BD74-BF6EDA787D57}.Release|x64.ActiveCfg = Release|x64
-		{7A352F3E-0C05-4EBA-BD74-BF6EDA787D57}.Release|x64.Build.0 = Release|x64
-		{7A352F3E-0C05-4EBA-BD74-BF6EDA787D57}.Release|x86.ActiveCfg = Release|Win32
-		{7A352F3E-0C05-4EBA-BD74-BF6EDA787D57}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {CDB5AD72-6CA8-40E4-919B-395F5B3617C9}
-	EndGlobalSection
-EndGlobal
