--- GraphicsUtilities/GraphicsUtilities.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.5.33530.505
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "GraphicsUtilities", "GraphicsUtilities.vcxproj", "{803A549A-4E75-4B80-BC1B-1822F2340C9F}"
-EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "GraphicsUtilitiesTest", "..\GraphicsUtilitiesTest\GraphicsUtilitiesTest.vcxproj", "{886D0A1C-3E1E-4111-B857-E8E7F62DE36D}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Release|x64 = Release|x64
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{803A549A-4E75-4B80-BC1B-1822F2340C9F}.Debug|x64.ActiveCfg = Debug|x64
-		{803A549A-4E75-4B80-BC1B-1822F2340C9F}.Debug|x64.Build.0 = Debug|x64
-		{803A549A-4E75-4B80-BC1B-1822F2340C9F}.Release|x64.ActiveCfg = Release|x64
-		{803A549A-4E75-4B80-BC1B-1822F2340C9F}.Release|x64.Build.0 = Release|x64
-		{886D0A1C-3E1E-4111-B857-E8E7F62DE36D}.Debug|x64.ActiveCfg = Debug|x64
-		{886D0A1C-3E1E-4111-B857-E8E7F62DE36D}.Debug|x64.Build.0 = Debug|x64
-		{886D0A1C-3E1E-4111-B857-E8E7F62DE36D}.Release|x64.ActiveCfg = Release|x64
-		{886D0A1C-3E1E-4111-B857-E8E7F62DE36D}.Release|x64.Build.0 = Release|x64
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {D27850E0-464C-4375-8208-BFDD0773E18D}
-	EndGlobalSection
-EndGlobal
