--- GraphicsUtilities/GraphicsUtilities.vcxproj
@@ -1,117 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<Project DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <ItemGroup Label="ProjectConfigurations">
-    <ProjectConfiguration Include="Debug|x64">
-      <Configuration>Debug</Configuration>
-      <Platform>x64</Platform>
-    </ProjectConfiguration>
-    <ProjectConfiguration Include="Release|x64">
-      <Configuration>Release</Configuration>
-      <Platform>x64</Platform>
-    </ProjectConfiguration>
-  </ItemGroup>
-  <ItemGroup>
-    <ClInclude Include="include\AvUiParser.h" />
-    <ClInclude Include="include\GraphicsUtilities.h" />
-    <ClInclude Include="src\core\components\component.h" />
-    <ClInclude Include="src\core\core.h" />
-    <ClInclude Include="src\core\logging\logging.h" />
-    <ClInclude Include="src\core\positioner\positioner.h" />
-    <ClInclude Include="src\core\renderer\renderer.h" />
-    <ClInclude Include="src\core\util\arena.h" />
-    <ClInclude Include="src\core\util\util.h" />
-    <ClInclude Include="src\parser\parser.h" />
-  </ItemGroup>
-  <ItemGroup>
-    <ClCompile Include="src\core\components\component.c" />
-    <ClCompile Include="src\core\core.c" />
-    <ClCompile Include="src\core\logging\logging.c" />
-    <ClCompile Include="src\core\renderer\vulkanCore\vulkanCore.c" />
-    <ClCompile Include="src\core\util\util.c" />
-    <ClCompile Include="src\parser\parser.c" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="include\AvGraphicsCore.h" />
-  </ItemGroup>
-  <PropertyGroup Label="Globals">
-    <VCProjectVersion>17.0</VCProjectVersion>
-    <ProjectGuid>{803a549a-4e75-4b80-bc1b-1822f2340c9f}</ProjectGuid>
-    <Keyword>Win32Proj</Keyword>
-    <WindowsTargetPlatformVersion>10.0</WindowsTargetPlatformVersion>
-  </PropertyGroup>
-  <Import Project="$(VCTargetsPath)\Microsoft.Cpp.Default.props" />
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|x64'" Label="Configuration">
-    <ConfigurationType>StaticLibrary</ConfigurationType>
-    <UseDebugLibraries>true</UseDebugLibraries>
-    <PlatformToolset>v143</PlatformToolset>
-  </PropertyGroup>
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|x64'" Label="Configuration">
-    <ConfigurationType>StaticLibrary</ConfigurationType>
-    <UseDebugLibraries>false</UseDebugLibraries>
-    <PlatformToolset>v143</PlatformToolset>
-  </PropertyGroup>
-  <Import Project="$(VCTargetsPath)\Microsoft.Cpp.props" />
-  <ImportGroup Label="ExtensionSettings">
-  </ImportGroup>
-  <ImportGroup Label="Shared">
-  </ImportGroup>
-  <ImportGroup Label="PropertySheets" Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">
-    <Import Project="$(UserRootDir)\Microsoft.Cpp.$(Platform).user.props" Condition="exists('$(UserRootDir)\Microsoft.Cpp.$(Platform).user.props')" Label="LocalAppDataPlatform" />
-  </ImportGroup>
-  <ImportGroup Label="PropertySheets" Condition="'$(Configuration)|$(Platform)'=='Release|x64'">
-    <Import Project="$(UserRootDir)\Microsoft.Cpp.$(Platform).user.props" Condition="exists('$(UserRootDir)\Microsoft.Cpp.$(Platform).user.props')" Label="LocalAppDataPlatform" />
-  </ImportGroup>
-  <PropertyGroup Label="UserMacros" />
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">
-    <TargetName>$(ProjectName)Debug</TargetName>
-    <OutDir>C:\SDK\CCR\lib\</OutDir>
-    <IntDir>x64\$(Configuration)\</IntDir>
-    <IncludePath>C:\SDK\CCR\include;C:\SDK\glfw-3.3.8.bin.WIN64\include;C:\SDK\VulkanSDK\1.3.243.0\Include;$(ProjectDir)include;$(IncludePath)</IncludePath>
-  </PropertyGroup>
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|x64'">
-    <OutDir>C:\SDK\CCR\lib\</OutDir>
-    <IntDir>x64\$(Configuration)\</IntDir>
-    <IncludePath>C:\SDK\CCR\include;C:\SDK\glfw-3.3.8.bin.WIN64\include;C:\SDK\VulkanSDK\1.3.243.0\Include;$(ProjectDir)include;$(IncludePath)</IncludePath>
-  </PropertyGroup>
-  <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">
-    <Lib>
-      <AdditionalLibraryDirectories>C:\SDK\CCR\lib;C:\SDK\glfw-3.3.8.bin.WIN64\lib-vc2022;C:\SDK\VulkanSDK\1.3.243.0\Lib;%(AdditionalLibraryDirectories)</AdditionalLibraryDirectories>
-      <LinkTimeCodeGeneration>false</LinkTimeCodeGeneration>
-      <AdditionalDependencies>glfw3.lib;vulkan-1.lib;MemoryUtilitiesDebug.lib;%(AdditionalDependencies)</AdditionalDependencies>
-    </Lib>
-    <PreBuildEvent>
-      <Command>IF exist "C:\SDK\CCR\include\$(ProjectName)" ( echo include for $(ProjectName) already exists, no action needed ) ELSE ( mkdir "C:\SDK\CCR\include\$(ProjectName)" &amp;&amp; echo include for $(ProjectName) created)</Command>
-    </PreBuildEvent>
-    <PostBuildEvent>
-      <Command>xcopy include C:\SDK\CCR\include\$(ProjectName) /e /y</Command>
-    </PostBuildEvent>
-    <ClCompile>
-      <LanguageStandard>stdcpp20</LanguageStandard>
-    </ClCompile>
-    <ClCompile>
-      <LanguageStandard_C>stdc11</LanguageStandard_C>
-    </ClCompile>
-  </ItemDefinitionGroup>
-  <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Release|x64'">
-    <Lib>
-      <AdditionalLibraryDirectories>C:\SDK\CCR\lib;C:\SDK\glfw-3.3.8.bin.WIN64\lib-vc2022;C:\SDK\VulkanSDK\1.3.243.0\Lib;%(AdditionalLibraryDirectories)</AdditionalLibraryDirectories>
-      <LinkTimeCodeGeneration>false</LinkTimeCodeGeneration>
-      <AdditionalDependencies>glfw3.lib;vulkan-1.lib;MemoryUtilitiesDebug.lib;%(AdditionalDependencies)</AdditionalDependencies>
-    </Lib>
-    <PreBuildEvent>
-      <Command>mkdir C:\SDK\CCR\include\$(ProjectName)</Command>
-    </PreBuildEvent>
-    <PostBuildEvent>
-      <Command>xcopy include C:\SDK\CCR\include\$(ProjectName) /e /y</Command>
-    </PostBuildEvent>
-    <ClCompile>
-      <LanguageStandard>stdcpp20</LanguageStandard>
-    </ClCompile>
-    <ClCompile>
-      <LanguageStandard_C>stdc11</LanguageStandard_C>
-    </ClCompile>
-  </ItemDefinitionGroup>
-  <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
-  <ImportGroup Label="ExtensionTargets">
-  </ImportGroup>
-</Project>
