--- GraphicsUtilities/design/design.drawio
@@ -1,1 +1,0 @@
-<mxfile host="Electron" modified="2023-05-08T11:52:40.599Z" agent="5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) draw.io/15.8.7 Chrome/91.0.4472.164 Electron/13.6.2 Safari/537.36" etag="e2UveFwB3v06PvtRjC0v" version="15.8.7" type="device"><diagram id="zrq5PWK3RwtQAuNaodZj" name="Page-1">7Vpbk6I4FP41Ps4Ud/Wx0ba3d3uqdtu5bD9tRYmQ7kCoEFrpX78JBgGDl50FZXpGrZKcnJDkO1/COScMzEm4uaMgDj4RD+KBoXmbgTkdGIaumSP+JyTZVuJY463Ap8iTSqVgjt5g0VJKU+TBpKbICMEMxXXhkkQRXLKaDFBK1nW1FcH1XmPgQ0UwXwKsSr8hjwVb6cgYlvLfIPKDomfdkfMLQaEsZ5IEwCPrisi8HZgTSgjbXoWbCcQCvAKXbbvZgdrdwCiM2DkNnH9e7Nj8/PUWBvAP14nnxH/+IO/yCnAqJzxFSYxBNk/pCiyhHDrLCjwoSSMPiltqA9NdB4jBeSwUzemaM4DLAhZiXtL55QphPCGY0Lyt6WhgrA+FnESsIl/lHy5PGCUvsFJjTIeOJjpSJ1uMHFIGNxWRnPwdJCFkNOMqOy5KQ0gm6oVh1qVdd7KgYlNHyoCkkr+7dYk2v5CA/wfwDQX8R8jBpVP4it4V9IbVN+hNBfqb1/soYSBqGXhNG4EcxrOB1zR72BXnLRV3x7kk7tYByneC/XVJb476RnpbAf8bisQz6f2AbhunQd+tgouA7iigzygI2yX6zWQ2s8dHsT1kjQ4wNzQV8+ElIR++e8gd7TTko0tCPnr/kI97xnJDfZY+gIykTIRIWcyDpn3wkzUKMYigBGouawRwywBhTzY3p/xZvHwpSm5AKHrj+qCwC6+mTIZqplbTmIuW8p4UJlznzwJ0fU/0CWxqig8gYcVoCMYgTtAiH59oGALqo8gljJFQKlXZ4cEVSDFrx9KWXTO002BovcnSun3E1LK3Rx6jgsjn86qs5Xp3w4butIbuzL3eAGaQRoBBV6yxROHXbqL/g3KqB3GzSAhOmbrSOdhMWasR2bKvYjopAhj5ES9iuBLNhLUQj8JvpDhEnifu7CZ8t0CR/5CrTa1S8iiRESLCm69wHmwHvCHkd3BjgiKWQ2W7/MfBm2gf7YHNxzrhZb0s859Qp3wn4V4pBShnEuT0XENBUZcSBhhY7BaPyrnjC/Y0E7O6iU8xb58K7e0xqvdyJ/I2v4zdvrGdqxtb9ZvmMUbsl7U7sPb4gtZeR/Phl7cNe/4wuV9u4F9PT19/b8j8zZBo9bPbWjHh2eY/aOvGyP+iS9s0FMNCz4eFB8hBCohPIoBvS6lb99VLnQdCYumWPUPGMukJgpSRuv9+0BdLSEqX8BigMp7gjqYPjyIvs+1iNkeXHYUYMPRaz+q3v4WqcdDHFCnYf3cU1KqfO6w7nrt81fWSsuY1WMoBpNnfsn1eeBIFvonI4nRTrZxmBx8NZ7N7fCa7i/2gL+weK+zmoVwCaU8JbuwlpsyG9LfelP7Wrc4obv0cFC+Ye5rieq8oXoy7QvFFJqJbcZgMsp4S3dJ7t5Wr7uWXe15epAh7DRvGCeRAEm8P91doI9DuDkrbOWPPaDpA6HDPULMuPeXh/pFXU+qsMS+td0dENYPxA4PXyLxRZ9ipCYH2sBPpeVt8mxL6Tv7hNT4FHoK1pP4i/3ZzdNUDzNUI4kfha9ND57LYqf5pT7FTeHd17CzV8fksUj9trndgaJphHzmmU/eBmaa19mrA3ktITkNWpikY+I6kDC+W7/VtD1/KtyPN238B</diagram></mxfile>
