--- GraphicsUtilities/src/core/components/component.h
@@ -1,6 +1,0 @@
-#pragma once
-#include "../core.h"
-
-typedef struct AvInterface_T {
-	int a;
-} AvInterface_T;
