--- GraphicsUtilities/src/core/positioner/positioner.h
@@ -1,2 +1,0 @@
-#pragma once
-#include "../core.h"
