--- GraphicsUtilities/src/core/util/arena.h
@@ -1,1 +1,0 @@
-#pragma once
