--- GraphicsUtilities/src/core/util/util.h
@@ -1,8 +1,0 @@
-#pragma once
-#include "../core.h"
-
-
-
-
-
-
