--- GraphicsUtilities/src/parser/parser.h
@@ -1,3 +1,0 @@
-#pragma once
-#include "../core/core.h"
-
