+++ src/core/components/component.h
@@ -1,0 +1,6 @@
+#pragma once
+#include "../core.h"
+
+typedef struct AvInterface_T {
+	int a;
+} AvInterface_T;
