+++ test/assets/testinterface.ui
@@ -1,0 +1,59 @@
+// Test interface
+// V1.0
+// 7-5-2023
+#once
+#include "./test.ui"
+#define SIZE 42
+
+Window(
+	width = 1280;
+	height = 720;
+	title = "test interface";
+	resizable = true;
+	fullscreen = false;
+	
+	backgroundColor = #000000FF;
+
+	root = $mainPanel;
+);
+
+@CustomButton = Rect(
+	label = Text(
+		text = "";
+		*font = Consolas;
+		*size = 12;
+	);
+	
+);
+
+@RedButton = CustomButton(
+	*color = #FF0000FF;
+);
+
+mainPanel = Rect(
+	layout = Grid(
+		numX = 2;
+		mumY = 2;
+	);
+);
+
+nestedPanel = Rect(
+	parent = $mainPanel;
+	x = 1;
+	y = 0;
+	color = #FFFFFFFF;
+	layout = absolute;
+);
+
+testButton[5] = RedButton(
+	parent = $nestedPanel;
+	x = 32;
+	y = 32;
+	width = 64;
+	height = 10;
+	label.text = "test";
+	label.font = Consolas;
+	label.size = 13;
+	onClick = $onClickFunction;
+);
+
